+++ src/pages/register-product-page.jsx
